Just use a char array :-(
Wanted to have the HttpRequest return a string or STL type but I
can't make it work for reasons as yet unknown. Suspect there is
weird concurrency stuff happening due to interrupts and a dual-core
CPU but I don't understand it well enough. So we just store the
body of the response in a fixed size buffer and then let the caller
turn that into a std::string.