Give up and use C strings
There is something really screwy going on (see
comment in HttpConnection constructor. Unless
things are _just right_ then the HTTP call
silently doesn't get sent to the server and I
don't understand lwIP (or indeed C++ in the
crazy embedded environment that it runs in on the
Pico) well enough to know why.