Download jpeg images using lwIP
Crucial in order to cause this not to run out of memory is to
acknowledge receipt of TCP packets with tcp_rcvd() and then free
the pbuf that contains the data that we just consumed. Without
this, it runs out of memory and fails to download the whole file.

The book actually doesn't tell you this (!) and it's really
hard to find any docs that tell you how to do it.